added files for KPI and fixed conflict

+++ dto/src/main/java/com/kairos/response/dto/web/wta/ChildCareDaysCheckWTATemplateDTO.java
@@ -15,7 +15,7 @@ public class ChildCareDaysCheckWTATemplateDTO extends WTABaseRuleTemplateDTO{
     private List<AgeRange> ageRange;
     private List<BigInteger> activityIds;
     private List<BigInteger> timeTypeIds = new ArrayList<>();
-    private List<Long> plannedTimeIds = new ArrayList<>();
+    private List<BigInteger> plannedTimeIds = new ArrayList<>();
     private LocalDate validationStartDate;
     private int numberOfWeeks;
     private boolean borrowLeave;
@@ -66,11 +66,11 @@ public class ChildCareDaysCheckWTATemplateDTO extends WTABaseRuleTemplateDTO{
         this.timeTypeIds = timeTypeIds;
     }
 
-    public List<Long> getPlannedTimeIds() {
+    public List<BigInteger> getPlannedTimeIds() {
         return plannedTimeIds;
     }
 
-    public void setPlannedTimeIds(List<Long> plannedTimeIds) {
+    public void setPlannedTimeIds(List<BigInteger> plannedTimeIds) {
         this.plannedTimeIds = plannedTimeIds;
     }
